code fully split and builds without complaints, still need to check loading to HW
