+++ Platforms/Android/PlatformUSB.cs
@@ -1,6 +1,8 @@
 ﻿using Android.Content;
 using Android.Hardware.Usb;
 using EnlightenMAUI.Models;
+using EnlightenMAUI.Common;
+using Microsoft.Maui;
 using Microsoft.Maui.Controls.PlatformConfiguration;
 using System;
 using System.Collections.Generic;
@@ -8,8 +10,281 @@ using System.Text;
 using static Android.Widget.GridLayout;
 using AndrApp = Android.App;
 
-namespace EnlightenMAUI.Platforms.Android
+namespace EnlightenMAUI.Platforms
 {
+    public  class USBWrapper
+    {
+        Logger logger = Logger.getInstance();
+        static UsbDeviceConnection udc;
+        static UsbDevice acc;
+        public int vid;
+        public int pid;
+
+        public const byte HOST_TO_DEVICE = 0x40;
+        public const byte DEVICE_TO_HOST = 0xc0;
+        Dictionary<Opcodes, byte> cmd = OpcodeHelper.getInstance().getDict();
+        HashSet<Opcodes> armInvertedRetvals = OpcodeHelper.getInstance().getArmInvertedRetvals();
+
+        public USBWrapper(UsbDeviceConnection udcI, UsbDevice accI)
+        {
+            udc = udcI;
+            acc = accI;
+            vid = accI.VendorId; 
+            pid = accI.ProductId;
+
+        }
+
+        public void disconnect()
+        {
+            udc.ReleaseInterface(acc.GetInterface(0));
+            udc.Close();
+        }
+
+        public bool connect()
+        {
+            bool ok = udc.SetConfiguration(acc.GetConfiguration(0));
+            if (ok)
+            {
+                logger.info("successfully set configuration");
+                ok = udc.ClaimInterface(acc.GetInterface(0), false);
+                if (ok)
+                {
+                    logger.info("successfully claimed interface");
+                    return true;
+                }
+            }
+
+            return false;
+        }
+
+        public async Task<int> bulkTransfer(byte[] spectrumBuff, int timeout)
+        {
+            logger.debug("buffer init with {0} in pix 0", spectrumBuff[0]);
+            int result = await udc.BulkTransferAsync(acc.GetInterface(0).GetEndpoint(0), spectrumBuff, spectrumBuff.Length, timeout);
+            logger.debug("buffer transfered with {0} in pix 0", spectrumBuff[0]);
+            return result;
+        }
+
+
+        /// <summary>
+        /// Execute a request-response control transaction using the given opcode.
+        /// </summary>
+        /// <param name="opcode">the opcode of the desired request</param>
+        /// <param name="len">the number of needed return bytes</param>
+        /// <param name="wIndex">an optional numeric argument used by some opcodes</param>
+        /// <param name="fullLen">the actual number of expected return bytes (not all needed)</param>
+        /// <remarks>not sure fullLen is actually required...testing</remarks>
+        /// <returns>the array of returned bytes (null on error)</returns>
+        internal async Task<byte[]> getCmdAsync(Opcodes opcode, int len, ushort wIndex = 0, int fullLen = 0, bool minRead = true)
+        {
+
+            int bytesToRead = Math.Max(len, fullLen);
+            if (minRead) // ARM should always read at least 8 bytes
+                bytesToRead = Math.Min(8, bytesToRead);
+            byte[] buf = new byte[bytesToRead];
+
+            logger.debug("about to send getCmd...");
+
+            int okI = await udc.ControlTransferAsync((UsbAddressing)DEVICE_TO_HOST, cmd[opcode], 0, wIndex, buf, bytesToRead, 100);
+
+            if (logger.debugEnabled())
+                logger.hexdump(buf, String.Format("getCmd: {0} (0x{1:x2}) index 0x{2:x4} ->", opcode.ToString(), cmd[opcode], wIndex));
+
+            // extract just the bytes we really needed
+            return await Task.Run(() => Util.truncateArray(buf, len));
+        }
+
+        internal byte[] getCmd(Opcodes opcode, int len, ushort wIndex = 0, int fullLen = 0, bool minRead = true)
+        {
+            int bytesToRead = Math.Max(len, fullLen);
+            if (minRead) // ARM should always read at least 8 bytes
+                bytesToRead = Math.Min(8, bytesToRead);
+            byte[] buf = new byte[bytesToRead];
+
+            logger.debug("about to send getCmd...");
+
+            int okI = udc.ControlTransfer((UsbAddressing)DEVICE_TO_HOST, cmd[opcode], 0, wIndex, buf, bytesToRead, 100);
+
+            if (logger.debugEnabled())
+                logger.hexdump(buf, String.Format("getCmd: {0} (0x{1:x2}) index 0x{2:x4} ->", opcode.ToString(), cmd[opcode], wIndex));
+
+            // extract just the bytes we really needed
+            return Util.truncateArray(buf, len);
+        }
+
+        /// <summary>
+        /// Execute a request-response transfer with a "second-tier" request.
+        /// </summary>
+        /// <param name="opcode">the wValue to send along with the "second-tier" command</param>
+        /// <param name="len">how many bytes of response are expected</param>
+        /// <returns>array of returned bytes (null on error)</returns>
+        /// 
+        internal async Task<byte[]> getCmd2Async(Opcodes opcode, int len, ushort wIndex = 0, int fakeBufferLengthARM = 0, bool minRead = true)
+        {
+            int bytesToRead = len;
+            if (minRead)
+                bytesToRead = Math.Max(bytesToRead, fakeBufferLengthARM);
+
+            byte[] buf = new byte[bytesToRead];
+
+            bool expectedSuccessResult = true;
+            if (minRead && armInvertedRetvals.Contains(opcode))
+                expectedSuccessResult = !expectedSuccessResult;
+
+            bool result = false;
+
+            int okI = await udc.ControlTransferAsync((UsbAddressing)DEVICE_TO_HOST, cmd[Opcodes.SECOND_TIER_COMMAND], cmd[opcode], wIndex, buf, bytesToRead, 100);
+
+            if (result != expectedSuccessResult && okI < len)
+            {
+                logger.error("getCmd2: failed to get SECOND_TIER_COMMAND {0} (0x{1:x4}) via DEVICE_TO_HOST ({2} of {3} bytes read, expected {4} got {5})",
+                    opcode.ToString(), cmd[opcode], okI, len, expectedSuccessResult, result);
+                logger.hexdump(buf, $"{opcode} result");
+                //return null;
+            }
+
+            logger.hexdump(buf, String.Format("getCmd2: {0} (0x{1:x2}) index 0x{2:x4} (result {3}, expected {4}) ->",
+                    opcode.ToString(), cmd[opcode], wIndex, result, expectedSuccessResult));
+
+            // extract just the bytes we really needed
+            return Util.truncateArray(buf, len);
+        }
+        internal byte[] getCmd2(Opcodes opcode, int len, ushort wIndex = 0, int fakeBufferLengthARM = 0, bool minRead = true)
+        {
+            int bytesToRead = len;
+            if (minRead)
+                bytesToRead = Math.Max(bytesToRead, fakeBufferLengthARM);
+
+            byte[] buf = new byte[bytesToRead];
+
+            bool expectedSuccessResult = true;
+            if (minRead && armInvertedRetvals.Contains(opcode))
+                expectedSuccessResult = !expectedSuccessResult;
+
+            bool result = false;
+
+            int okI = udc.ControlTransfer((UsbAddressing)DEVICE_TO_HOST, cmd[Opcodes.SECOND_TIER_COMMAND], cmd[opcode], wIndex, buf, bytesToRead, 100);
+
+            if (result != expectedSuccessResult || okI < len)
+            {
+                logger.error("getCmd2: failed to get SECOND_TIER_COMMAND {0} (0x{1:x4}) via DEVICE_TO_HOST ({2} of {3} bytes read, expected {4} got {5})",
+                    opcode.ToString(), cmd[opcode], okI, len, expectedSuccessResult, result);
+                logger.hexdump(buf, $"{opcode} result");
+                return null;
+            }
+
+            if (logger.debugEnabled())
+                logger.hexdump(buf, String.Format("getCmd2: {0} (0x{1:x2}) index 0x{2:x4} (result {3}, expected {4}) ->",
+                    opcode.ToString(), cmd[opcode], wIndex, result, expectedSuccessResult));
+
+            // extract just the bytes we really needed
+            return Util.truncateArray(buf, len);
+        }
+
+        /// <summary>
+        /// send a single control transfer command (response not checked)
+        /// </summary>
+        /// <param name="opcode">the desired command</param>
+        /// <param name="wValue">an optional secondary argument used by most commands</param>
+        /// <param name="wIndex">an optional tertiary argument used by some commands</param>
+        /// <param name="buf">a data buffer used by some commands</param>
+        /// <returns>true on success, false on error</returns>
+        /// <todo>should support return code checking...most cmd opcodes return a success/failure byte</todo>
+        /// 
+        internal async Task<bool> sendCmdAsync(Opcodes opcode, ushort wValue = 0, ushort wIndex = 0, byte[] buf = null, bool minRead = true)
+        {
+            if ((minRead) && (buf is null || buf.Length < 8))
+                buf = new byte[8];
+
+            ushort wLength = (ushort)((buf is null) ? 0 : buf.Length);
+
+            bool? expectedSuccessResult = true;
+            if (minRead)
+            {
+                if (opcode != Opcodes.SECOND_TIER_COMMAND)
+                    expectedSuccessResult = armInvertedRetvals.Contains(opcode);
+                else
+                    expectedSuccessResult = null; // no easy way to know, as we don't pass wValue as enum (MZ: whut?)
+            }
+
+            int okI = await udc.ControlTransferAsync((UsbAddressing)HOST_TO_DEVICE, cmd[opcode], wValue, wIndex, buf, wLength, 100);
+
+            if (opcode == Opcodes.ACQUIRE_SPECTRUM)
+            {
+                logger.info("sendCmd: failed to send {0} (0x{1:x2}) (wValue 0x{2:x4}, wIndex 0x{3:x4}, wLength 0x{4:x4}) (received {5}, expected {6})",
+                    opcode.ToString(), cmd[opcode], wValue, wIndex, wLength, okI, expectedSuccessResult);
+            }
+
+            if (expectedSuccessResult != null && okI < 0)
+            {
+                logger.error("sendCmd: failed to send {0} (0x{1:x2}) (wValue 0x{2:x4}, wIndex 0x{3:x4}, wLength 0x{4:x4}) (received {5}, expected {6})",
+                    opcode.ToString(), cmd[opcode], wValue, wIndex, wLength, okI, expectedSuccessResult);
+                return false;
+            }
+
+            return true;
+        }
+        internal bool sendCmd(Opcodes opcode, ushort wValue = 0, ushort wIndex = 0, byte[] buf = null, bool minRead = true)
+        {
+            if ((minRead) && (buf is null || buf.Length < 8))
+                buf = new byte[8];
+
+            ushort wLength = (ushort)((buf is null) ? 0 : buf.Length);
+
+            bool? expectedSuccessResult = true;
+            if (minRead)
+            {
+                if (opcode != Opcodes.SECOND_TIER_COMMAND)
+                    expectedSuccessResult = armInvertedRetvals.Contains(opcode);
+                else
+                    expectedSuccessResult = null; // no easy way to know, as we don't pass wValue as enum (MZ: whut?)
+            }
+
+            int okI = udc.ControlTransfer((UsbAddressing)HOST_TO_DEVICE, cmd[opcode], wValue, wIndex, buf, wLength, 100);
+
+            if (expectedSuccessResult != null && okI < 0)
+            {
+                logger.error("sendCmd: failed to send {0} (0x{1:x2}) (wValue 0x{2:x4}, wIndex 0x{3:x4}, wLength 0x{4:x4}) (received {5}, expected {6})",
+                    opcode.ToString(), cmd[opcode], wValue, wIndex, wLength, okI, expectedSuccessResult);
+                return false;
+            }
+
+            return true;
+        }
+
+        /// <summary>
+        /// send a single 2nd-tier control transfer command (response not checked)
+        /// </summary>
+        /// <param name="opcode">the desired command</param>
+        /// <param name="wIndex">an optional secondary argument used by some 2nd-tier commands</param>
+        /// <param name="buf">a data buffer used by some commands</param>
+        /// <returns>true on success, false on error</returns>
+        /// <todo>should support return code checking...most cmd opcodes return a success/failure byte</todo>
+        internal async Task<bool> sendCmd2Async(Opcodes opcode, ushort wIndex = 0, byte[] buf = null, bool minRead = true)
+        {
+            if ((minRead) && (buf is null || buf.Length < 8))
+                buf = new byte[8];
+
+            ushort wLength = (ushort)((buf is null) ? 0 : buf.Length);
+
+            int okI = await udc.ControlTransferAsync((UsbAddressing)HOST_TO_DEVICE, cmd[Opcodes.SECOND_TIER_COMMAND], cmd[opcode], wIndex, buf, wLength, 100);
+
+            return true;
+        }
+        internal bool sendCmd2(Opcodes opcode, ushort wIndex = 0, byte[] buf = null, bool minRead = true)
+        {
+            if ((minRead) && (buf is null || buf.Length < 8))
+                buf = new byte[8];
+
+            ushort wLength = (ushort)((buf is null) ? 0 : buf.Length);
+
+            int okI = udc.ControlTransfer((UsbAddressing)HOST_TO_DEVICE, cmd[Opcodes.SECOND_TIER_COMMAND], cmd[opcode], wIndex, buf, wLength, 100);
+
+            return true;
+        }
+
+    }
+
     internal class PlatformUSB
     {
         static UsbDevice device;
@@ -96,7 +371,8 @@ namespace EnlightenMAUI.Platforms.Android
                     if (udc != null)
                     {
                         logger.info("successfully opened device");
-                        USBSpectrometer usbSpectrometer = new USBSpectrometer(udc, device);
+                        USBWrapper passthrough = new USBWrapper(udc, device);
+                        USBSpectrometer usbSpectrometer = new USBSpectrometer(passthrough);
                         spec = usbSpectrometer;
 
                         bool ok = await (spec as USBSpectrometer).initAsync();
