+++ Platforms/iOS/AppDelegate.cs
@@ -1,0 +1,13 @@
+﻿using Foundation;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace EnlightenMAUI
+{ 
+    [Register("AppDelegate")]
+    public class AppDelegate : MauiUIApplicationDelegate
+    {
+        protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
+    }
+}
